% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_numbers.R
\name{add_magic}
\alias{add_magic}
\title{Add a Magic Number}
\usage{
add_magic(val, desc = NULL, list_name = "magic")
}
\arguments{
\item{val}{value to save; its variable name is used as the record name}

\item{desc}{optional str description of the value}

\item{list_name}{str name of the list to update, in the calling
environment}
}
\value{
side effect only: updates \code{list_name} in the calling environment
}
\description{
Add a single value (with an optional description) to a
running named list, using the variable name passed as \code{val} as the
record name. Meant to avoid retyping the name when building up a list for
\code{\link[=save_magic]{save_magic()}}, e.g. \code{add_magic(n_indicators, "Total indicators")}
adds a \code{n_indicators} entry. Creates \code{list_name} in the calling
environment if it doesn't already exist.
}
