% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_numbers.R
\name{create_magic_sty_file}
\alias{create_magic_sty_file}
\title{Create Magic Sty File}
\usage{
create_magic_sty_file(input_files, output_file)
}
\arguments{
\item{input_files}{string path(s) of yaml files to combine}

\item{output_file}{str path to output \code{.sty} file}
}
\value{
side effect only: save yaml file
}
\description{
Turn one or more \code{.yaml} files into a single \code{magic_numbers.sty}
file for use with latex.
}
